pyproject.toml
[project]
name = "ai-code-reviewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.5.3",
    "cryptography>=46.0.2",
    "fastapi>=0.118.0",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-google>=0.1.1",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "uvicorn[standard]>=0.37.0",
]
